refactor: modularize face landmarker
